pyproject.toml
[project]
name = "reykit"
version = "1.2.56"
description = "Kit method set."
readme = "README.md"
requires-python = ">=3.12"
license = {file = "LICENSE"}
keywords = ["rey", "reyxbo", "kit", "tool"]
authors = [
    {name = "Rey", email = "reyxbo@163.com"}
]
dependencies = [
    "aiohttp",
    "apscheduler",
    "bcrypt",
    "captcha",
    "concurrent-log-handler",
    "filetype",
    "pandas",
    "pyzbar",
    "pdfplumber",
    "psutil",
    "pymem",
    "python-docx",
    "qrcode",
    "requests",
    "requests_cache",
    "reydb",
    "varname",
    "Pillow",
    "PyJWT"
]

[project.urls]
Homepage = "https://www.reyxbo.com"
Repository = "https://github.com/reyxbo/reykit-py.git"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
